Refactor startup checks into domain-specific modules (#692)
